fix: print a debug message if lua_ls is not found
If you're using this config somewhere lua_ls is not installed. It
currently shows the warning and grab focus and attention. Meaning
we need to manually "escape" to get out of the message before we
can use Neovim.
This commit changes this message to a debug message, so it still
shows but do not grab focus. You can simply start using neovim.